Removed Invocable.InvocableCompletableFuture
