+++ jetty-core/jetty-util/src/main/java/org/eclipse/jetty/util/thread/Invocable.java
@@ -13,16 +13,6 @@
 
 package org.eclipse.jetty.util.thread;
 
-import java.util.Objects;
-import java.util.concurrent.CompletionStage;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-import java.util.function.BiConsumer;
-import java.util.function.BiFunction;
-import java.util.function.Consumer;
-import java.util.function.Function;
-
 /**
  * <p>A task (typically either a {@link Runnable} or {@link Callable}
  * that declares how it will behave when invoked:</p>
@@ -267,157 +257,5 @@ public interface Invocable
             }
         }
         return result;
-    }
-
-    /**
-     * An extension of {@link java.util.concurrent.CompletableFuture} that is an {@link Invocable}.
-     * The {@link InvocationType} is the type passed in construction (default NON_BLOCKING).
-     * Methods on {@link java.util.concurrent.CompletableFuture} that may act in contradiction to the passed
-     * {@link InvocationType} are extended to throw {@link IllegalStateException} in those circumstances.
-     * <p>
-     * Counterintuitively, if the blocking APIs like {@link #get()} are to be used, the {@link InvocationType}
-     * should be {@link InvocationType#NON_BLOCKING}, as the wake-up callbacks used will not block.
-     *
-     * @param <V> The type of the result
-     * @deprecated This class in only used for deprecated usages of CompletableFuture
-     */
-    @Deprecated(forRemoval = true, since = "12.0.15")
-    class InvocableCompletableFuture<V> extends java.util.concurrent.CompletableFuture<V> implements Invocable
-    {
-        private final InvocationType _invocationType;
-
-        public InvocableCompletableFuture(InvocationType invocationType)
-        {
-            _invocationType = Objects.requireNonNull(invocationType);
-            if (_invocationType == InvocationType.EITHER)
-                throw new IllegalArgumentException("EITHER is not supported");
-        }
-
-        @Override
-        public InvocationType getInvocationType()
-        {
-            return _invocationType;
-        }
-
-        @Override
-        public V get() throws InterruptedException, ExecutionException
-        {
-            if (getInvocationType() == InvocationType.BLOCKING && !isDone())
-                throw new IllegalStateException("Must be NON_BLOCKING or completed");
-            return super.get();
-        }
-
-        @Override
-        public V get(long timeout, TimeUnit unit) throws InterruptedException, ExecutionException, TimeoutException
-        {
-            if (getInvocationType() == InvocationType.BLOCKING && !isDone())
-                throw new IllegalStateException("Must be NON_BLOCKING or completed");
-            return super.get(timeout, unit);
-        }
-
-        @Override
-        public V join()
-        {
-            if (!isDone() && getInvocationType() == InvocationType.BLOCKING && !isDone())
-                throw new IllegalStateException("Must be NON_BLOCKING or completed");
-            return super.join();
-        }
-
-        @Override
-        public java.util.concurrent.CompletableFuture<Void> acceptEither(CompletionStage<? extends V> other, Consumer<? super V> action)
-        {
-            if (!isDone() && getInvocationType() != combineTypes(getInvocationType(), Invocable.getInvocationType(other), Invocable.getInvocationType(action)))
-                throw new IllegalStateException("Bad invocation type when not completed");
-
-            return super.acceptEither(other, action);
-        }
-
-        @Override
-        public <U> java.util.concurrent.CompletableFuture<U> applyToEither(CompletionStage<? extends V> other, Function<? super V, U> fn)
-        {
-            if (!isDone() && getInvocationType() != combineTypes(getInvocationType(), Invocable.getInvocationType(other), Invocable.getInvocationType(fn)))
-                throw new IllegalStateException("Bad invocation type when not completed");
-            return super.applyToEither(other, fn);
-        }
-
-        @Override
-        public <U> java.util.concurrent.CompletableFuture<U> handle(BiFunction<? super V, Throwable, ? extends U> fn)
-        {
-            if (!isDone() && getInvocationType() != combine(getInvocationType(), Invocable.getInvocationType(fn)))
-                throw new IllegalStateException("Bad invocation type when not completed");
-            return super.handle(fn);
-        }
-
-        @Override
-        public java.util.concurrent.CompletableFuture<Void> runAfterBoth(CompletionStage<?> other, Runnable action)
-        {
-            if (!isDone() && getInvocationType() != combine(getInvocationType(), Invocable.getInvocationType(action)))
-                throw new IllegalStateException("Bad invocation type when not completed");
-            return super.runAfterBoth(other, action);
-        }
-
-        @Override
-        public java.util.concurrent.CompletableFuture<Void> runAfterEither(CompletionStage<?> other, Runnable action)
-        {
-            if (!isDone() && getInvocationType() != combineTypes(getInvocationType(), Invocable.getInvocationType(other), Invocable.getInvocationType(action)))
-                throw new IllegalStateException("Bad invocation type when not completed");
-            return super.runAfterEither(other, action);
-        }
-
-        @Override
-        public java.util.concurrent.CompletableFuture<Void> thenAccept(Consumer<? super V> action)
-        {
-            if (!isDone() && getInvocationType() != combine(getInvocationType(), Invocable.getInvocationType(action)))
-                throw new IllegalStateException("Bad invocation type when not completed");
-            return super.thenAccept(action);
-        }
-
-        @Override
-        public <U> java.util.concurrent.CompletableFuture<Void> thenAcceptBoth(CompletionStage<? extends U> other, BiConsumer<? super V, ? super U> action)
-        {
-            if (!isDone() && getInvocationType() != combineTypes(getInvocationType(), Invocable.getInvocationType(other), Invocable.getInvocationType(action)))
-                throw new IllegalStateException("Bad invocation type when not completed");
-            return super.thenAcceptBoth(other, action);
-        }
-
-        @Override
-        public <U> java.util.concurrent.CompletableFuture<U> thenApply(Function<? super V, ? extends U> fn)
-        {
-            if (!isDone() && getInvocationType() != combine(getInvocationType(), Invocable.getInvocationType(fn)))
-                throw new IllegalStateException("Bad invocation type when not completed");
-            return super.thenApply(fn);
-        }
-
-        @Override
-        public <U, V1> java.util.concurrent.CompletableFuture<V1> thenCombine(CompletionStage<? extends U> other, BiFunction<? super V, ? super U, ? extends V1> fn)
-        {
-            if (!isDone() && getInvocationType() != combineTypes(getInvocationType(), Invocable.getInvocationType(other), Invocable.getInvocationType(fn)))
-                throw new IllegalStateException("Bad invocation type when not completed");
-            return super.thenCombine(other, fn);
-        }
-
-        @Override
-        public <U> java.util.concurrent.CompletableFuture<U> thenCompose(Function<? super V, ? extends CompletionStage<U>> fn)
-        {
-            if (!isDone() && getInvocationType() != combine(getInvocationType(), Invocable.getInvocationType(fn)))
-                throw new IllegalStateException("Bad invocation type when not completed");
-            return super.thenCompose(fn);
-        }
-
-        @Override
-        public java.util.concurrent.CompletableFuture<Void> thenRun(Runnable action)
-        {
-            if (!isDone() && getInvocationType() != combine(getInvocationType(), Invocable.getInvocationType(action)))
-                throw new IllegalStateException("Bad invocation type when not completed");
-            return super.thenRun(action);
-        }
-
-        @Override
-        public java.util.concurrent.CompletableFuture<V> whenComplete(BiConsumer<? super V, ? super Throwable> action)
-        {
-            if (!isDone() && getInvocationType() != combine(getInvocationType(), Invocable.getInvocationType(action)))
-                throw new IllegalStateException("Bad invocation type when not completed");
-            return super.whenComplete(action);
-        }
     }
 }
