+++ jetty-ee10/jetty-ee10-servlet/src/main/java/org/eclipse/jetty/ee10/servlet/ServletMultiPartFormData.java
@@ -200,7 +200,7 @@ public class ServletMultiPartFormData
                         .maxSize(config.getMaxRequestSize())
                         .build();
 
-                    futureServletParts = new Invocable.InvocableCompletableFuture<>(invocationType);
+                    futureServletParts = new CompletableFuture<>();
                     CompletableFuture<Parts> futureConvertParts = futureServletParts;
 
                     Promise.Invocable<MultiPartFormData.Parts> onParts = new Promise.Invocable<>()
